Add eventstream doc
