+++ docs/source/user/cookbook/06_eventstream_01.py
@@ -1,0 +1,13 @@
+from messagebus import Event, Field, Metadata
+
+
+class BookRegistered(Event):
+    id: str = Field(...)
+    isbn: str = Field(...)
+    title: str = Field(...)
+    author: str = Field(...)
+    metadata: Metadata = Metadata(
+        name="register_book",
+        schema_version=1,
+        published=True,
+    )
