+++ docs/source/user/cookbook/06_eventstream_03.py
