+++ docs/source/user/cookbook/06_eventstream_04.py
