+++ src/messagebus/service/_sync/registry.py
@@ -15,9 +15,9 @@ from messagebus.typing import (
     SyncCommandHandler,
     SyncEventHandler,
     SyncMessageHandler,
-    TSyncUow,
     TCommand,
     TEvent,
+    TSyncUow,
 )
 
 from .unit_of_work import SyncUnitOfWorkTransaction, TRepositories
