+++ tests/_sync/test_eventstore.py
@@ -2,12 +2,12 @@ from messagebus.domain.model import Metadata
 from messagebus.service._sync.registry import SyncMessageBus
 from messagebus.service._sync.unit_of_work import SyncUnitOfWorkTransaction
 from tests._sync.conftest import (
-    SyncDummyUnitOfWorkWithEvents,
-    SyncEventstreamTransport,
     DummyCommand,
     DummyEvent,
     DummyModel,
     Repositories,
+    SyncDummyUnitOfWorkWithEvents,
+    SyncEventstreamTransport,
 )
