+++ tests/_sync/test_unit_of_work.py
@@ -8,7 +8,7 @@ from messagebus.service._sync.unit_of_work import (
     TransactionError,
     TransactionStatus,
 )
-from tests._sync.conftest import SyncDummyUnitOfWork, DummyModel
+from tests._sync.conftest import DummyModel, SyncDummyUnitOfWork
 
 
 class FooCreated(Event):
@@ -20,9 +20,7 @@ class BarCreated(Event):
     metadata: Metadata = Metadata(name="bar_created", schema_version=1, published=True)
 
 
-def test_collect_new_events(
-    uow: SyncDummyUnitOfWork, foo_factory: Type[DummyModel]
-):
+def test_collect_new_events(uow: SyncDummyUnitOfWork, foo_factory: Type[DummyModel]):
     foo = foo_factory(id="1", counter=0)
     foo.messages.append(FooCreated(id="1"))
     bar = foo_factory(id="1", counter=0)
